refactor to use goview
