+++ pkg/view/router.go
@@ -1,11 +1,11 @@
 package view
 
 import (
+	"github.com/foolin/goview"
 	"github.com/go-chi/chi/v5"
+	"github.com/scratchdata/scratchdata/pkg/config"
 	"github.com/scratchdata/scratchdata/pkg/storage/database"
-	"github.com/scratchdata/scratchdata/pkg/view/templates"
-	"html/template"
-	"io/fs"
+	"github.com/scratchdata/scratchdata/templates"
 	"net/http"
 )
 
@@ -13,18 +13,24 @@ type Model struct {
 	Email string
 }
 
-func withLayout(f fs.FS, t ...string) (*template.Template, error) {
-	base := []string{"layout/base.html"}
-	base = append(base, t...)
-	return template.ParseFS(f, base...)
+func embeddedFH(config goview.Config, tmpl string) (string, error) {
+	bytes, err := templates.Templates.ReadFile(tmpl + config.Extension)
+	return string(bytes), err
 }
 
-func New(auth func(h http.Handler) http.Handler) (*chi.Mux, error) {
+func New(c config.DashboardConfig, auth func(h http.Handler) http.Handler) (*chi.Mux, error) {
 	r := chi.NewRouter()
 	r.Use(auth)
 
-	// TODO breadchris setup reloading of templates
-	f := templates.Templates
+	gv := goview.New(goview.Config{
+		Root:         "templates",
+		Extension:    ".html",
+		Master:       "layout/base",
+		DisableCache: true,
+	})
+	if !c.LiveReload {
+		gv.SetFileHandler(embeddedFH)
+	}
 
 	loadModel := func(r *http.Request) Model {
 		userAny := r.Context().Value("user")
@@ -37,45 +43,29 @@ func New(auth func(h http.Handler) http.Handler) (*chi.Mux, error) {
 		}
 	}
 
-	dash, err := withLayout(f, "pages/index.html")
-	if err != nil {
-		return nil, err
-	}
 	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
-		err := dash.Execute(w, loadModel(r))
+		err := gv.Render(w, http.StatusOK, "pages/index", loadModel(r))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
 	})
 
-	conns, err := withLayout(f, "pages/connections/index.html")
-	if err != nil {
-		return nil, err
-	}
 	r.Get("/connections", func(w http.ResponseWriter, r *http.Request) {
-		err := conns.Execute(w, loadModel(r))
+		err := gv.Render(w, http.StatusOK, "pages/connections/index", loadModel(r))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
 	})
 
-	connsNew, err := withLayout(f, "pages/connections/new.html")
-	if err != nil {
-		return nil, err
-	}
 	r.Get("/connections/new", func(w http.ResponseWriter, r *http.Request) {
-		err := connsNew.Execute(w, loadModel(r))
+		err := gv.Render(w, http.StatusOK, "pages/connections/new", loadModel(r))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
 	})
 
-	keys, err := withLayout(f, "pages/keys/index.html")
-	if err != nil {
-		return nil, err
-	}
 	r.Get("/keys", func(w http.ResponseWriter, r *http.Request) {
-		err := keys.Execute(w, loadModel(r))
+		err := gv.Render(w, http.StatusOK, "pages/keys/index", loadModel(r))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
